Separated api into views and project routes and deleted unused routes
