+++ pcb_anodet_server/project_routes/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+api_blueprint = Blueprint("project_routes_api_blueprint", __name__, url_prefix="")
