+++ pcb_anodet_server/templates/edit/take_photo.html
@@ -6,7 +6,7 @@
 <div class="row content">
   <div class="col-sm-4 col-sm-offset-4 live-stream"> 
     <h5>Live feed</h5>
-      <img id="video_feed" style="width:300px; height:300px;" src="{{ url_for('api_blueprint.video_feed') }}">
+      <img id="video_feed" style="width:300px; height:300px;" src="{{ url_for('project_routes_api_blueprint.video_feed') }}">
   </div>
   <div class="col-sm-4 text-left"></div>
 </div>
@@ -24,7 +24,7 @@
       <span id="save-info-txt">	&nbsp;</span>
     </div>
     <hr>
-    <a id="go-next-btn" class="btn btn-block btn-success" href="{{ url_for('api_blueprint.view_imagesv') }}">Go to next step</a>
+    <a id="go-next-btn" class="btn btn-block btn-success" href="{{ url_for('views_blueprint.view_images_view') }}">Go to next step</a>
   </div>
 </div>
