+++ pcb_anodet_server/templates/edit/view_images.html
@@ -11,9 +11,9 @@
 <div class="row content">
   <div class="col-sm-12">
     <br>
-    <a id="take-more-photos-btn" class="btn btn-block btn-primary" href="{{ url_for('api_blueprint.take_photov') }}">Take more photos</a>
+    <a id="take-more-photos-btn" class="btn btn-block btn-primary" href="{{ url_for('views_blueprint.take_photo_view') }}">Take more photos</a>
     <hr>
-    <a id="go-next-btn" class="btn btn-block btn-success" href="{{ url_for('api_blueprint.train_projectv') }}">Go to next step</a>
+    <a id="go-next-btn" class="btn btn-block btn-success" href="{{ url_for('views_blueprint.train_project_view') }}">Go to next step</a>
   </div>
 </div>
