+++ pcb_anodet_server/templates/predict.html
@@ -15,7 +15,7 @@
   <div class="col-sm-6"> 
     <h5>Live feed</h5>
     <div style="height: 400px; max-height: 400px;">
-      <img id="video_feed" style="max-height: 100%; max-width: 100%; min-height: 400px;" src="{{ url_for('api_blueprint.video_feed') }}"></img>
+      <img id="video_feed" style="max-height: 100%; max-width: 100%; min-height: 400px;" src="{{ url_for('project_routes_api_blueprint.video_feed') }}"></img>
     </div>
   </div>
