+++ pcb_anodet_server/views/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+api_blueprint = Blueprint("views_blueprint", __name__, url_prefix="")
